refactor: collection view private 처리
